refactor: split sim_codegen/ram.rs into submodule
Extracts gen_ram (~343 LoC) into src/sim_codegen/ram.rs. Mirrors
#73 pattern for fsm/pipeline. No behavior change.

cargo test: 158/158 pass, unchanged.
